First Misery Of gabriel

+++ MKproject/Management/ClientProfile/UCBundlePackage.cs
@@ -524,7 +524,7 @@ namespace MKproject.Management
         {
           
            //sql
-            int lastClientId = ParentFormClientMan.Client.ClientId;
+            int lastClientId = (int)ParentFormClientMan.Client.ClientId;
             if (BundleId != null)
             {
                 ProjectToSQL.InsertToClientBalance(lastClientId, (int)BundleId, BundleType.ToString());
@@ -547,7 +547,7 @@ namespace MKproject.Management
             {
                 action = "Purchased the " + BundleDescription + " Package.";
             }
-            ClassBackOffice backOffice = new ClassBackOffice(ParentFormClientMan.Client.ClientId, action, ActionsEnum.Purchases, LOGIN.Employee.EmployeeId, (int)InsertedRow["ID"], null, null, null, null, DateTime.Now);
+            ClassBackOffice backOffice = new ClassBackOffice((int)ParentFormClientMan.Client.ClientId, action, ActionsEnum.Purchases, LOGIN.Employee.EmployeeId, (int)InsertedRow["ID"], null, null, null, null, DateTime.Now);
             backOffice.InsertToArchiveSQL();
             
 
@@ -649,9 +649,9 @@ namespace MKproject.Management
             int UpdatedSessionLeft = SessionDaysLeft-1;
             DateTime Date = DateTime.Now;
             ProjectToSQL.UpdateClientBalanceOnEditingSessions(ID, UpdatedSessionLeft, null, null, false);//lieanno this function onlykermel el package sessiosns                   
-            ClassClient.UpdateClientCheckInSQL(ParentFormClientMan.Client.ClientId, Date);
-            ProjectToSQL.InsertToClientAttendance(ParentFormClientMan.Client.ClientId);
-            ClassBackOffice backOffice = new ClassBackOffice(ParentFormClientMan.Client.ClientId, "Completed a session.", ActionsEnum.SessionDone, LOGIN.Employee.EmployeeId, Id, null, SQLToProject.GetLAstInsertedAttendance(), null, null, Date);
+            ClassClient.UpdateClientCheckInSQL((int)ParentFormClientMan.Client.ClientId, Date);
+            ProjectToSQL.InsertToClientAttendance((int)ParentFormClientMan.Client.ClientId);
+            ClassBackOffice backOffice = new ClassBackOffice((int)ParentFormClientMan.Client.ClientId, "Completed a session.", ActionsEnum.SessionDone, LOGIN.Employee.EmployeeId, Id, null, SQLToProject.GetLAstInsertedAttendance(), null, null, Date);
             backOffice.InsertToArchiveSQL();
 
             //design
@@ -700,7 +700,7 @@ namespace MKproject.Management
             double EntetityAmount = (double)foundRow["balance"];
 
 
-            Payment payment = new Payment(ParentFormClientMan.Client.ClientId, EntetityAmount, ParentFormClientMan.RetrievingSpecificRowsInDt(false, Id), ParentFormClientMan);
+            Payment payment = new Payment((int)ParentFormClientMan.Client.ClientId, EntetityAmount, ParentFormClientMan.RetrievingSpecificRowsInDt(false, Id), ParentFormClientMan);
             payment.ClientManagementProfileParentForm = this.ParentFormClientMan;
             payment.Show();
         }
