+++ MKproject/Management/Payment/Payment.cs
@@ -417,7 +417,7 @@ namespace MKproject.Management
                     //logic ended
 
                     //SQL
-                    backOffice = new ClassBackOffice(ClientManagementProfileParentForm.Client.ClientId, "Paid " + Currency.Symbol + AmoutPerRow + " for the " + BackOfficeCatName + BackOfficeCatType + ".", ActionsEnum.Payments, LOGIN.Employee.EmployeeId, ClientBalanceId, AmoutPerRow, null, null, null, Date);
+                    backOffice = new ClassBackOffice((int)ClientManagementProfileParentForm.Client.ClientId, "Paid " + Currency.Symbol + AmoutPerRow + " for the " + BackOfficeCatName + BackOfficeCatType + ".", ActionsEnum.Payments, LOGIN.Employee.EmployeeId, ClientBalanceId, AmoutPerRow, null, null, null, Date);
                     backOffice.InsertToArchiveSQL();
 
 
@@ -470,7 +470,7 @@ namespace MKproject.Management
                     //logic ended
 
                     //Sql
-                    backOffice = new ClassBackOffice(ClientManagementProfileParentForm.Client.ClientId, "Paid " + Currency.Symbol + AmountPaid + " for the " + BackOfficeCatName + " " + BackOfficeCatType + ".", ActionsEnum.Payments, LOGIN.Employee.EmployeeId, ClientBalanceId, AmountPaid, null, null, null, Date);
+                    backOffice = new ClassBackOffice((int)ClientManagementProfileParentForm.Client.ClientId, "Paid " + Currency.Symbol + AmountPaid + " for the " + BackOfficeCatName + " " + BackOfficeCatType + ".", ActionsEnum.Payments, LOGIN.Employee.EmployeeId, ClientBalanceId, AmountPaid, null, null, null, Date);
                     backOffice.InsertToArchiveSQL();
 
                     //Design
